created WeatherApp class
